geom_segment2

geom_segment2 support aes(subset) via setup_data

Usage

geom_segment2(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  lineend = "butt",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  nudge_x = 0,
  arrow = NULL,
  arrow.fill = NULL,
  ...
)

Arguments

mapping

Set of aesthetic mappings, defaults to NULL

data

A layer specific dataset - only needed if you want to override the plot defaults.

stat

Name of stat to modify data.

position

The position adjustment to use for overlapping points on this layer.

lineend

Line end style, one of butt (default), round and square.

na.rm

If "FALSE" (default), missing values are removed with a warning. If "TRUE", missing values are silently removed, logical.

show.legend

Whether to show legend, logical.

inherit.aes

Whether to inherit aesthetic mappings, logical, defaults to "TRUE".

nudge_x

adjust the horizontal position of the segments.

arrow

specification for arrow heads, as created by arrow().

arrow.fill

fill color to usse for the arrow head (if closed). NULL means use colour aesthetic.

...

additional parameter

Value

add segment layer

Details

'geom_segment2' is a modified version of geom_segment, with subset aesthetic supported

See also

Author

Guangchuang Yu